feat: introduce conftest
