+++ tests/conftest.py
@@ -1,0 +1,37 @@
+from unittest.mock import patch
+
+import pytest
+from fastapi.testclient import TestClient
+
+import aidial_analytics_realtime.app as app
+from tests.mocks import InfluxWriterMock, TestTopicModel
+from tests.utils.client import Client
+
+
+@pytest.fixture
+def mock_uuid4():
+    counter = 0
+
+    def side_effect() -> str:
+        nonlocal counter
+        counter += 1
+        return f"pseudo-uuid-{counter}"
+
+    with patch(
+        "aidial_analytics_realtime.analytics.uuid4", side_effect=side_effect
+    ):
+        yield
+
+
+@pytest.fixture
+def influx():
+    return InfluxWriterMock()
+
+
+@pytest.fixture
+def client(influx) -> Client:
+    app.app.dependency_overrides[app.InfluxWriterAsync] = lambda: influx  # type: ignore
+    app.app.dependency_overrides[app.TopicModel] = lambda: TestTopicModel()
+    return Client(
+        http_client=TestClient(app.app, raise_server_exceptions=False)
+    )
