+++ tests/endpoints/test_rate.py
@@ -1,25 +1,7 @@
 import json
 
-import pytest
-from fastapi.testclient import TestClient
-
-import aidial_analytics_realtime.app as app
-from tests.mocks import InfluxWriterMock, TestTopicModel
+from tests.mocks import InfluxWriterMock
 from tests.utils.client import Client
-
-
-@pytest.fixture
-def influx():
-    return InfluxWriterMock()
-
-
-@pytest.fixture
-def client(influx) -> Client:
-    app.app.dependency_overrides[app.InfluxWriterAsync] = lambda: influx  # type: ignore
-    app.app.dependency_overrides[app.TopicModel] = lambda: TestTopicModel()
-    return Client(
-        http_client=TestClient(app.app, raise_server_exceptions=False)
-    )
 
 
 def test_rate_request(client: Client, influx: InfluxWriterMock):
